Update hours difference function
Removed unecessary float conversion

+++ learn-to-program-the-fundamentals-python/Assignment-1/_11e8aef37a1877652d7b666a7f0c27f7_a1.py
@@ -14,7 +14,7 @@ def seconds_difference(time_1, time_2):
     0.0
     """
     
-    return float((time_2 - time_1)/3600)
+    return (time_2 - time_1)/3600
 
 def hours_difference(time_1, time_2):
     """ (number, number) -> float
